refactor e2e code
